requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-threadlocals==0.10
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
Pillow==8.1.2
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.3
webencodings==0.5.1
XlsxWriter==1.3.7
xmltodict==0.12.0
